8-8

+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\DataTables\SubCategoryDataTable;
 use App\Models\Category;
 use App\Models\SubCategory;
+use App\Models\ChildCategory;
 use Str;
 
 class SubCategoryController extends Controller
@@ -96,6 +97,10 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+        if($childCategory > 0){
+            return response(['status' => 'error', 'message' => 'This category contain sub category, unable to delete!']);
+        }
         $subCategory->delete();
         return redirect()->route('admin.sub-category.index');
     }
